XRay mobile quantized model
Summary:
This is going to allow to experiment with various training from scratch / fine tunning technics. The code itself for the new model is not intended to be used as is. Instead one could  train a full precision model first. Then add quantization for the last layer, then for the next one and so on.

In my experiments I tried getting a pretrained model and then quantizing all inception layers with 4 bits. This restored original accuracy after several dozen iterations

Also in this diff I added a common prefix to the model checkpoint + added this prefix to git / hg ignore.
And also some extra logs which are usefull to quickly see how things changed right after enabling quantization

Differential Revision: D4672026

fbshipit-source-id: b022c8ccf11dd8a2af1a7b2e92673483bc741a11