fix CUDNN layer weight size calculation for multiple layers
Summary: CuDNN LSTM weights were incorrectly sized for layers > 0: there was assumption that the input size to middle layers is same as for the first layer, but actually the middle layer will get input from a layer below, which will have dimension equal to the output dimension (hidden dimension). This worked fine when input_dim and hidden_dim were equal, as are the default params for lstm_benchmark.

Reviewed By: salexspb

Differential Revision: D4922824

fbshipit-source-id: 3ed05529dcb0a4e66ad440084a55df1c5932fd33