caffe2: smart_tensor_printer
Summary:
As per request moving elsewhere and using the Dispatcher. The reason
why I didn't put it into tensor.h is because the dispatcher lives in operator.h
and operator.h includes tensor.h. I also didn't want to do any codemods. If
this turns out to be useful it can be changed. Also the name is not super great
but the TensorPrinter is already taken so that's what first came to mind.

Reviewed By: dzhulgakov

Differential Revision: D4893325

fbshipit-source-id: 7d4e56c4e57164c3cd3748f4a705a4ffe6b932d9