fix curand odd-sized workaround
Summary: Ran into illegal memory access errors when running MSRAFill on an odd-sized tensor.  curand only supports even-sized fills.  To workaround this limitation, we fill the last entry of the tensor manually and use curand for what remains.  In this line, the intent is to get the n-1 th element of the tensor.  r is already a T*, so we should not be multiplying by sizeof(T) to get the n-1 th element.

Differential Revision: D4961306

fbshipit-source-id: 587f2945abf025e28f573482a4828c09e6ae771b