MultiRNNCell (Caffe2)
Summary: Add Python support for arbitrary (unidirectional) recurrent networks with MultiRNNCell abstraction. Since the combined step net for all layers is created at one time (in method _apply), this may be optimizable as-is. LSTM() function is extended to accept a list of numbers of units for the dim_out argument, producing a multi-layer LSTM in that case.

Reviewed By: salexspb

Differential Revision: D4965001

fbshipit-source-id: 39c069468d5b40bf803503cf62046a479ca83cbb