Fix residual connections
Summary:
This diff fixes gradient computation of residual connections for a training network constructed with MultiRNNCell.

It addresses a logic bug in _prepare_output() and _prepare_output_sequence() by keeping track internally of which layers have consecutive residual connections before the output, and then reconstructing the final residual output by (re-)preparing the output of each of those layers and then combining them with a Sum operation. This also involves keeping track of which states contribute toward the reconstruction of the final sequence output so that outputs_with_grads can be correctly passed to apply_over_sequence().

Differential Revision: D5300520

fbshipit-source-id: f37d800c909e631175de7045abe192351cc11c41